fix: migrate code to updated bevy
* includes updates for bevy-inspector-egui api
