+++ Cargo.toml
@@ -8,13 +8,14 @@ default = []
 debug = ["board_plugin/debug", "bevy-inspector-egui"]
 
 [dependencies]
-bevy = "0.11"
+bevy.workspace = true
 board_plugin = { path = "board_plugin" }
-
-# Hierarchy inspector debug
-bevy-inspector-egui = { version = "0.18.3", optional = true }
+bevy-inspector-egui = { version = "0.19.0", optional = true }
 
 [workspace]
 members = [
     "board_plugin"
 ]
+
+[workspace.dependencies]
+bevy = "0.11"
