+++ board_plugin/Cargo.toml
@@ -8,8 +8,8 @@ default = []
 debug = ["colored", "bevy-inspector-egui"]
 
 [dependencies]
-bevy = "0.11"
+bevy.workspace = true
 serde = "1.0"
 rand = "0.8"
 colored = { version = "2.0", optional = true }
-bevy-inspector-egui = { version = "0.18.3", optional = true }
+bevy-inspector-egui = { version = "0.19.0", optional = true }
