+++ board_plugin/src/components/coordinates.rs
@@ -1,12 +1,18 @@
-use bevy::prelude::Component;
+use bevy::prelude::*;
 use std::fmt::{self, Display, Formatter};
 use std::ops::{Add, Sub};
 
-#[cfg_attr(feature = "debug", derive(bevy_inspector_egui::Inspectable))]
+#[cfg_attr(feature = "debug", derive(Reflect))]
 #[derive(Debug, Default, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash, Component)]
 pub struct Coordinates {
     pub x: u16,
     pub y: u16,
+}
+
+impl Coordinates {
+    pub fn new(x: u16, y: u16) -> Self {
+        Self { x, y }
+    }
 }
 
 impl Add for Coordinates {
@@ -17,6 +23,17 @@ impl Add for Coordinates {
             x: self.x + rhs.x,
             y: self.y + rhs.y,
         }
+    }
+}
+
+impl Add<(i8, i8)> for Coordinates {
+    type Output = Self;
+
+    fn add(self, (x, y): (i8, i8)) -> Self::Output {
+        let x = ((self.x as i16) + x as i16) as u16;
+        let y = ((self.y as i16) + y as i16) as u16;
+
+        Self { x, y }
     }
 }
