+++ rust-toolchain.toml
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "nightly"
-# channel = "stable"
+# channel = "nightly"
+channel = "stable"
